requirements.txt
streamlit==1.28.1
pandas==2.1.3
numpy==1.24.3
requests==2.31.0
beautifulsoup4==4.12.2
yfinance==0.2.28
plotly==5.17.0
openai==1.3.7
python-dotenv==1.0.0
newsapi-python==0.2.6
textblob==0.17.1
nltk==3.8.1
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
transformers==4.35.2
torch==2.1.1
langchain==0.0.340
langchain-openai==0.0.2
